# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: StayHub V2 API
  version: 1.0.0
  description: Static OpenAPI contract aligned with the currently implemented StayHub V2 controllers.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v2/bookings/book:
    post:
      tags:
        - Bookings
      summary: Create a new accommodation reservation
      description: |
        Creates a reservation for an accommodation.

        Business notes:
        - reservations must be created at least 72 hours before check-in,
        - overlapping active reservations are rejected,
        - the response includes the deposit amount, bank account number and payment deadline.
      operationId: bookAccommodation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
      responses:
        '201':
          description: Reservation successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateReservationResponse'
        '400':
          description: Invalid request or reservation policy violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Accommodation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Accommodation not available for selected dates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v2/bookings/{reservationId}:
    get:
      tags:
        - Bookings
      summary: Get reservation detail by ID
      description: Returns the full detail of a reservation for the guest owner or accommodation host.
      operationId: getReservationById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reservationId
          required: true
          description: Reservation identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Reservation retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveReservationResponse'
        '400':
          description: Invalid reservation ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Bookings
      summary: Update an active reservation date range
      description: |
        Updates the check-in and check-out dates of an active reservation owned by the authenticated guest.

        If the new check-in is within the next 72 hours, the deposit must already be paid.
      operationId: updateReservation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reservationId
          required: true
          description: Reservation identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReservationRequest'
      responses:
        '200':
          description: Reservation updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveReservationResponse'
        '400':
          description: Invalid request or reservation policy violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Only the guest can update the reservation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Dates unavailable or deposit not paid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v2/bookings/{reservationId}/cancel:
    patch:
      tags:
        - Bookings
      summary: Cancel an active reservation
      description: Cancels an active reservation owned by the authenticated guest. Cancellation is allowed only at least 48 hours before check-in.
      operationId: cancelReservation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reservationId
          required: true
          description: Reservation identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Reservation cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveReservationResponse'
        '400':
          description: Reservation cannot be cancelled by policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Only the guest can cancel the reservation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v2/bookings/{reservationId}/deposit-paid:
    patch:
      tags:
        - Bookings
      summary: Mark reservation deposit as paid
      description: Marks the advance deposit as paid for an active reservation owned by the authenticated guest.
      operationId: markDepositAsPaid
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reservationId
          required: true
          description: Reservation identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Deposit marked as paid successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveReservationResponse'
        '400':
          description: Reservation is not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Only the guest can mark the deposit as paid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v2/bookings/my-reservations:
    get:
      tags:
        - Bookings
      summary: Get paginated list of my reservations
      description: |
        Returns paginated reservations for the authenticated user.

        If `scope` is omitted, legacy behavior is preserved:
        - users with role `HOST` receive host reservations,
        - users without role `HOST` receive guest reservations.
      operationId: getMyReservations
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          required: false
          description: Zero-based page number.
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: scope
          required: false
          description: Optional reservation filter scope.
          schema:
            type: string
            enum:
              - host
              - guest
              - all
      responses:
        '200':
          description: Reservations retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveReservationSummaryPage'
        '400':
          description: Invalid page or scope parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      description: Standard error payload returned by the API.
      properties:
        message:
          type: string
          example: Invalid login attempt detected
        code:
          type: integer
          example: 401
      required:
        - message
        - code
...
    ReservationStatus:
      type: string
      enum:
        - ACTIVE
        - CANCELLED
        - COMPLETED
...
    CreateReservationRequest:
      type: object
      properties:
        accommodationId:
          type: integer
          format: int64
          example: 1
        startDate:
          type: string
          format: date-time
          example: '2025-06-01T15:00:00'
        endDate:
          type: string
          format: date-time
          example: '2025-06-05T11:00:00'
      required:
        - accommodationId
        - startDate
        - endDate
    UpdateReservationRequest:
      type: object
      properties:
        startDate:
          type: string
          format: date-time
          example: '2025-06-03T15:00:00'
        endDate:
          type: string
          format: date-time
          example: '2025-06-07T11:00:00'
      required:
        - startDate
        - endDate
    CreateReservationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12345
        startDate:
          type: string
          format: date-time
          example: '2025-06-01T15:00:00'
        endDate:
          type: string
          format: date-time
          example: '2025-06-05T11:00:00'
        totalPrice:
          type: number
          format: double
          example: 1000.00
        currency:
          type: string
          example: COP
        status:
          $ref: '#/components/schemas/ReservationStatus'
        accommodationId:
          type: integer
          format: int64
          example: 1
        accommodationTitle:
          type: string
          example: Beachfront Villa with Pool
        userId:
          type: integer
          format: int64
          example: 678
        depositAmount:
          type: number
          format: double
          example: 200.00
        bankAccountNumber:
          type: string
          example: '3001234567890'
        paymentDeadline:
          type: string
          format: date-time
          example: '2025-04-18T23:59:59'
      required:
        - id
        - startDate
        - endDate
        - totalPrice
        - currency
        - status
        - accommodationId
        - accommodationTitle
        - userId
        - depositAmount
        - bankAccountNumber
        - paymentDeadline
    RetrieveReservationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12345
        startDate:
          type: string
          format: date-time
          example: '2025-06-01T15:00:00'
        endDate:
          type: string
          format: date-time
          example: '2025-06-05T11:00:00'
        accommodationId:
          type: integer
          format: int64
          example: 1
        accommodationTitle:
          type: string
          example: Beachfront Villa with Pool
        accommodationCity:
          type: string
          example: Medellin
        guestId:
          type: integer
          format: int64
          example: 678
        guestEmail:
          type: string
          format: email
          example: johndoe@gmail.com
        totalPrice:
          type: number
          format: double
          example: 1250000
        currency:
          type: string
          example: COP
        depositAmount:
          type: number
          format: double
          example: 250000
        depositPaid:
          type: boolean
          example: true
        paymentDeadline:
          type: string
          format: date-time
          example: '2025-06-04T23:59:59'
        status:
          $ref: '#/components/schemas/ReservationStatus'
        createdAt:
          type: string
          format: date-time
          example: '2025-06-01T15:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2025-06-01T15:00:00'
      required:
        - id
        - startDate
        - endDate
        - accommodationId
        - accommodationTitle
        - accommodationCity
        - guestId
        - guestEmail
        - totalPrice
        - currency
        - depositAmount
        - depositPaid
        - paymentDeadline
        - status
        - createdAt
        - updatedAt
    RetrieveReservationSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12345
        accommodationId:
          type: integer
          format: int64
          example: 1
        accommodationTitle:
          type: string
          example: Beachfront Villa with Pool
        startDate:
          type: string
          format: date-time
          example: '2025-06-01T15:00:00'
        endDate:
          type: string
          format: date-time
          example: '2025-06-05T11:00:00'
        totalPrice:
          type: number
          format: double
          example: 1250000
        currency:
          type: string
          example: COP
        status:
          $ref: '#/components/schemas/ReservationStatus'
      required:
        - id
        - accommodationId
        - accommodationTitle
        - startDate
        - endDate
        - totalPrice
        - currency
        - status
    RetrieveReservationSummaryPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/RetrieveReservationSummaryResponse'
        pageable:
          type: object
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        numberOfElements:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
      required:
        - content
...